refactor(e-sign-svc): contract uplift + Phase 35 cleanup
- Phase 35 lifted 12 DTOs to dristi-common/contract/esign/
  (ESignRequest, ESignResponse, SignDocRequest, CoordinateRequest,
  CoordinateResponse, CoordinateCriteria, StorageResponse + 5
  transitive payload types: Coordinate, ESignParameter,
  ESignXmlForm, SignDocParameter, File)
- Phase 35 false-positive cleanup: dropped orphan
  `import org.pucar.dristi.common.contract.esign.File;` from 3 main
  + 1 test file that want `java.io.File`. File DTO stays in
  dristi-common because StorageResponse.files: List<File> uses it.
  Matches feedback_lifted_dto_shadowing memory.
- Removed provisional internal/web/models/package-info.java
  (@NamedInterface("contract") moved to dristi-common/contract/esign).
- 0 REST->direct conversions in C2 (rest_calls.txt empty for e-sign-svc).
  Inbound REST from bail-bond + digitalizeddocuments to esign endpoints
  deferred to follow-up PR (no EsignApi exposed yet).

Verified: full reactor `mvn clean test` -> 1149 tests pass across
domain-identity-access (135), domain-case-lifecycle (906),
domain-integration (70), domain-payments (36), dristi-app (2 — Bean
NameCollisionTest + ModuleStructureTest). All 8 pipeline gates PASS.
