+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/esign/internal/service/PdfEmbedder.java
@@ -12,8 +12,8 @@ import org.pucar.dristi.integration.esign.internal.config.Configuration;
 import org.pucar.dristi.integration.esign.internal.util.ByteArrayMultipartFile;
 import org.pucar.dristi.integration.esign.internal.util.FileStoreUtil;
 import org.pucar.dristi.integration.esign.internal.util.TextLocationFinder;
-import org.pucar.dristi.integration.esign.internal.web.models.Coordinate;
-import org.pucar.dristi.integration.esign.internal.web.models.ESignParameter;
+import org.pucar.dristi.common.contract.esign.Coordinate;
+import org.pucar.dristi.common.contract.esign.ESignParameter;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
