--- dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/esign/internal/web/models/SignDocRequest.java
@@ -1,27 +1,0 @@
-package org.pucar.dristi.integration.esign.internal.web.models;
-
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.Valid;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.egov.common.contract.request.RequestInfo;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class SignDocRequest {
-
-    @JsonProperty("RequestInfo")
-    @Valid
-    private RequestInfo requestInfo = null;
-
-    @JsonProperty("SignDocParameter")
-    @Valid
-    private SignDocParameter eSignParameter = null;
-
-
-}
