+++ dristi-monolith/domain-integration/src/test/java/org/pucar/dristi/integration/esign/internal/web/controllers/EsignApiControllerTest.java
@@ -3,10 +3,10 @@ package org.pucar.dristi.integration.esign.internal.web.controllers;
 import jakarta.servlet.http.HttpServletRequest;
 import org.pucar.dristi.integration.esign.internal.service.ESignService;
 import org.pucar.dristi.common.util.ResponseInfoFactory;
-import org.pucar.dristi.integration.esign.internal.web.models.ESignRequest;
-import org.pucar.dristi.integration.esign.internal.web.models.ESignResponse;
-import org.pucar.dristi.integration.esign.internal.web.models.ESignXmlForm;
-import org.pucar.dristi.integration.esign.internal.web.models.SignDocRequest;
+import org.pucar.dristi.common.contract.esign.ESignRequest;
+import org.pucar.dristi.common.contract.esign.ESignResponse;
+import org.pucar.dristi.common.contract.esign.ESignXmlForm;
+import org.pucar.dristi.common.contract.esign.SignDocRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
