+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/esign/StorageResponse.java
@@ -1,4 +1,5 @@
-package org.pucar.dristi.integration.esign.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.esign;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
